FEAT: tour api의 통신 부분과 가공 부분 분리 (#114)
